finished: task 6 TODO: task 7, 9

+++ DrawSVG/src/software_renderer.h
@@ -80,7 +80,7 @@ class SoftwareRendererImp : public SoftwareRenderer {
 
  protected:
 
-  uint8_t * buffer;
+  float * buffer;
 
  private:
